refactor: split thulp-cli main.rs into modules (1849→385 lines)
Pure structural refactor — zero logic changes:
- src/output.rs: OutputFormat enum + Output helper (47 LOC)
- src/commands/config.rs: ConfigCommands + handle_init + handle_config (308 LOC)
- src/commands/skill.rs: SkillCommands + all skill handlers (470 LOC)
- src/commands/tools.rs: ToolCommands + handle_run + demo + validate (364 LOC)
- src/commands/convert.rs: ConvertCommands + OpenAPI conversion (181 LOC)
- src/commands/mcp.rs: McpCommands + MCP handlers (134 LOC, feature-gated)
- src/main.rs: Cli/Commands enum + dispatch + tests (385 LOC)

32 tests pass (26 unit + 6 integration).
