+++ crates/thulp-cli/src/commands/mod.rs
@@ -1,3 +1,7 @@
 pub mod config;
+pub mod convert;
 pub mod skill;
 pub mod tools;
+
+#[cfg(feature = "mcp")]
+pub mod mcp;
